(Huge CL) Update Cobalt to use m70 Chromium repos
This is part of the project to update Cobalt's Chromium repos. Now
everything can compile and Cobalt is usable. A few tests are still
failing.

There is no need to go through everything, but please review these files:
base/wrap_main_starboard.h
browser/trace_manager.h(deleted)
debug/remote/debug_web_server.*
dom/html_image_element.*
h5vcc/dial/*
script/callback_function.h
math/r_tree_base.h
math/r_tree_unittest.cc
media/fetcher_buffered_data_source.*
network/cookie_jar_impl.*
network/network_delegate.*
netowrk/network_module.*
network/net_log_logger.*
network/url_request_context.*
network/persistent_cookie_store.*
network/cobalt_net_log.*
trace_event/scoped_event_parser_trace.*
script/union_type.h
loader/cobalt_url_response_string_writer.*
loader/net_fetcher.*
websocket/*(deleted and modified many files)
xhr/xml_http_request.cc

b/27798606

+++ cobalt/media_capture/encoders/linear16_audio_encoder.cc
@@ -14,9 +14,10 @@
 
 #include "cobalt/media_capture/encoders/linear16_audio_encoder.h"
 
+#include <algorithm>
 #include <string>
 
-#include "base/string_util.h"
+#include "base/strings/string_util.h"
 
 namespace {
 // See https://tools.ietf.org/html/rfc2586 for MIME type
@@ -45,7 +46,7 @@ bool Linear16AudioEncoder::IsLinear16MIMEType(
 void Linear16AudioEncoder::Encode(const ShellAudioBus& audio_bus,
                                   base::TimeTicks reference_time) {
   DCHECK_EQ(audio_bus.sample_type(), ShellAudioBus::kInt16);
-  DCHECK_EQ(audio_bus.channels(), 1);
+  DCHECK_EQ(audio_bus.channels(), size_t(1));
   auto data = audio_bus.interleaved_data();
   size_t data_size = audio_bus.GetSampleSizeInBytes() * audio_bus.frames();
   PushDataToAllListeners(data, data_size, reference_time);
